requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
pdfplumber
python-docx
openai
python-dotenv
scikit-learn
sentence-transformers
numpy
